Added parallel module visitors (#9)
* WIP sequential and parallel module visitors

* Added sequential/parallel visitor executor

* Use GCC8, disable visit tests & execution policies

* Fixed warning on shadow local variable

* Reduce GCC warnings

* Reduce GCC warnings

* Disable warning for useless cast inside SFML

* Disable useless cast warning

* Add execution header

* Experimental execution header

* Support parallel execution with MSVC compiler

* Improve test coverage

+++ nexus-lib/src/sfml/window.h
@@ -11,7 +11,7 @@ struct window {
 		: m_window{} {
 	}
 
-	window(uint16_t width, uint16_t height, const std::string& title = "unnamed")
+	window(uint32_t width, uint32_t height, const std::string& title = "unnamed")
 		: m_window(sf::VideoMode(width, height), title) {
 	}
